test: move 'remove-if-exists' to helper directory, which we'll use for other test helpers as well
