+++ test/jobs/github-event/marketplace_purchase/cancelled.js
@@ -1,7 +1,7 @@
 const { test, tearDown } = require('tap')
 const dbs = require('../../../../lib/dbs')
 const worker = require('../../../../jobs/github-event/marketplace_purchase/cancelled')
-const removeIfExists = require('../../../remove-if-exists.js')
+const removeIfExists = require('../../../helpers/remove-if-exists.js')
 
 test('marketplace canceled', async t => {
   t.test('change entry in payments database to `free`', async t => {
